requirements.txt
# Real-time Voice Interaction Agent Dependencies

# OpenAI API
openai>=1.12.0

# Audio processing
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0

# Environment variables
python-dotenv>=1.0.0

# Async support (usually included in Python 3.7+)
asyncio

# Optional: Edge TTS for ultra-fast speech synthesis (recommended)
edge-tts>=6.1.0